refactor: renamed from DeviceDetailsSDM630.py to DeviceDetailsSDM630Widget.py
