+++ pyqt/widgets/DeviceDetailsSDM630Widget.py
